--- sfx_utils/__init__.py
@@ -1,4 +1,0 @@
-from .psana_interface import *
-from .radial_utils import *
-from .geom_opt import *
-from .ag_behenate import *
